+++ lib/license_finder/package_managers/maven.rb
@@ -5,11 +5,14 @@ module LicenseFinder
   class Maven < PackageManager
     def initialize(options={})
       super
+      @ignore_groups = options[:ignore_groups]
       @include_groups = options[:maven_include_groups]
     end
 
     def current_packages
       command = "#{package_management_command} license:download-licenses"
+      command += " -Dlicense.excludedScopes=#{@ignore_groups.to_a.join(',')}" if @ignore_groups and !@ignore_groups.empty?
+
       output, success = Dir.chdir(project_path) { capture(command) }
       raise "Command '#{command}' failed to execute: #{output}" unless success